Add files via upload
Problem Statement :

There have always been queries of people regarding various topics to which answers are provided manually by people. In order to make this process fast the FAQs can be answer using the latest AI technologies where answers can be automatically generated according to the questions. Also for every question asked it can be stored and the list of questions can be increased to provide better mapping of question and answers. This process can reduce lot of work pressure for both the consumer and the government. Also providing quick responses and suggestions can help in the better economical market growth.
Summary: There have always been queries of people regarding various topics to which answers are provided manually by people. In order to make this process fast the FAQs can be answer using the latest AI technologies where answers can be automatically generated according to the questions. Objective: We need you to develop an AI based Chatbot (mobile/web app) to create answers to queries based on FAQs and more FAQs getting automatically added including categorization. You can use machine learning algorithms to increase the dataset of questions and answers and ensure that the mapping of questions and answers remain maintained every time there is an increase in the datase